Extract coop TUI review rendering
Committed-By-Agent: codex
